+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/nik-rev/cargo-reedme/compare/v0.4.1...HEAD
 
+- Documentation improvements
+
 ## [v0.4.1] - 2026-02-27
 
 [v0.4.1]: https://github.com/nik-rev/cargo-reedme/compare/v0.4.0...v0.4.1